rename views.py -> api.py
