+++ backend/projects/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from projects import views
+from backend.projects import api
 
 urlpatterns = [
-    path('api/projects/', views.ProjectView.as_view({'get': 'list'})),
+    path('api/projects/', api.ProjectView.as_view({'get': 'list'})),
 ]
